+++ contracts/multisig/test/asset_proxy_owner.ts
@@ -1,4 +1,3 @@
-import { artifacts as proxyArtifacts, MixinAuthorizableContract } from '@0x/contracts-asset-proxy';
 import {
     chaiSetup,
     constants,
@@ -26,6 +25,7 @@ import {
     AssetProxyOwnerExecutionFailureEventArgs,
     AssetProxyOwnerSubmissionEventArgs,
     AssetProxyOwnerWrapper,
+    MixinAuthorizableContract,
     TestAssetProxyOwnerContract,
 } from '../src';
 
@@ -58,12 +58,12 @@ describe('AssetProxyOwner', () => {
         notOwner = accounts[3];
         const initialOwner = accounts[0];
         erc20Proxy = await MixinAuthorizableContract.deployFrom0xArtifactAsync(
-            proxyArtifacts.MixinAuthorizable,
+            artifacts.MixinAuthorizable,
             provider,
             txDefaults,
         );
         erc721Proxy = await MixinAuthorizableContract.deployFrom0xArtifactAsync(
-            proxyArtifacts.MixinAuthorizable,
+            artifacts.MixinAuthorizable,
             provider,
             txDefaults,
         );