+++ contracts/utils/src/wrappers/index.ts
@@ -1,2 +1,0 @@
-export * from '../../generated-wrappers/test_constants';
-export * from '../../generated-wrappers/test_lib_bytes';