+++ packages/types/src/index.ts
@@ -45,9 +45,14 @@ export interface SignedOrder extends Order {
  * ZeroExTransaction for use with 0x Exchange executeTransaction
  */
 export interface ZeroExTransaction {
+    verifyingContractAddress: string;
     salt: BigNumber;
     signerAddress: string;
     data: string;
+}
+
+export interface SignedZeroExTransaction extends ZeroExTransaction {
+    signature: string;
 }
 
 /**
@@ -684,4 +689,34 @@ export interface DutchAuctionDetails {
     endAmount: BigNumber;
     currentAmount: BigNumber;
     currentTimeSeconds: BigNumber;
+}
+
+export interface PackageJSONConfig {
+    postpublish?: {
+        assets?: string[];
+        docOmitExports?: string[];
+        dockerHubRepo?: string;
+    };
+    'abis:comment'?: string;
+    abis?: string;
+    ignoreDependencyVersions?: string;
+    ignoreDependencyVersionsForPackage?: string;
+}
+
+export interface PackageJSON {
+    private?: boolean;
+    version: string;
+    name: string;
+    main?: string;
+    scripts?: { [command: string]: string };
+    config?: PackageJSONConfig;
+    dependencies?: { [dependencyName: string]: string };
+    devDependencies?: { [dependencyName: string]: string };
+    workspaces?: string[];
+}
+
+export interface EIP712DomainWithDefaultSchema {
+    name?: string;
+    version?: string;
+    verifyingContractAddress: string;
 }