+++ contracts/extensions/CHANGELOG.md
@@ -5,6 +5,18 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v2.0.3 - _February 7, 2019_
+
+    * Dependencies updated
+
+## v2.0.2 - _February 7, 2019_
+
+    * Fake publish to enable pinning
+
+## v2.0.1 - _February 6, 2019_
+
+    * Dependencies updated
+
 ## v2.0.0 - _February 5, 2019_
 
     * Upgrade the bignumber.js to v8.0.2 (#1517)