+++ contracts/test-utils/CHANGELOG.md
@@ -5,9 +5,22 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v3.0.3 - _February 7, 2019_
+
+    * Dependencies updated
+
+## v3.0.2 - _February 7, 2019_
+
+    * Fake publish to enable pinning
+
+## v3.0.1 - _February 6, 2019_
+
+    * Dependencies updated
+
 ## v3.0.0 - _February 5, 2019_
 
     * Upgrade the bignumber.js to v8.0.2 (#1517)
+    * Import `ZeroExTransaction` type instead of using type defined in this package (#1576)
 
 ## v2.0.1 - _January 17, 2019_