+++ packages/utils/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v4.0.3 - _February 7, 2019_
+
+    * Dependencies updated
+
+## v4.0.2 - _February 6, 2019_
+
+    * ABI Decode NULL for all data types (#1587)
+
 ## v4.0.1 - _February 5, 2019_
 
     * ABI Decode NULL as False (#1582)